+++ conversion/__init__.py
@@ -1,0 +1,18 @@
+"""
+conversion
+========
+A Python package for visualising and comparing the results of
+numerical simulations from pregnant and non-pregnant
+uterine smooth muscle cells.
+
+Modules:
+- metrics: Comparison metrics.
+- utils: General utilities for handling data and file operations.
+- constants: Constants used in the project.
+- plots: Plotting functions.
+- Means2023: Pregnant uterine cell model
+- Roesler2024: Non-pregnant uterine cell model
+
+"""
+
+__version__ = "0.1.0"
